test: long column name

+++ libs/output-mapping/tests/Writer/TableDefinitionV2Test.php
@@ -127,7 +127,7 @@ class TableDefinitionV2Test extends AbstractTestCase
                     'name' => 'Id',
                     'data_type' => [
                         'base' => [
-                            'type' => 'NUMERIC',
+                            'type' => 'INTEGER',
                         ],
                     ],
                 ],
